pyproject.toml
[tool.poetry]
name = "flask-starter"
version = "0.1.0"
description = "A starter repo for flask applications"
authors = ["Andrew Shoemaker <andrew.shoemaker9@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "1.15.1"
bcrypt = "4.3.0"
blinker = "1.9.0"
cffi = "1.17.1"
click = "8.1.8"
cryptography = "44.0.2"
dnspython = "2.7.0"
email-validator = "2.2.0"
flask = "3.1.0"
flask-login = "0.6.3"
flask-mail = "0.10.0"
flask-migrate = "4.1.0"
flask-sqlalchemy = "3.1.1"
flask-wtf = "1.2.2"
greenlet = "3.1.1"
gunicorn = "23.0.0"
idna = "3.10"
iniconfig = "2.1.0"
itsdangerous = "2.2.0"
jinja2 = "3.1.6"
mako = "1.3.9"
markupsafe = "3.0.2"
packaging = "24.2"
pluggy = "1.5.0"
pycparser = "2.22"
pyjwt = "2.10.1"
pytest = "8.3.5"
pytest-flask = "1.3.0"
python-dotenv = "1.0.1"
sqlalchemy = "2.0.39"
typing-extensions = "4.12.2"
werkzeug = "3.1.3"
wtforms = "3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
